+++ ChangeLog.md
@@ -58,6 +58,7 @@
 * `boost::iostreams` is no longer a build dependency
 * `boost::filesystem` is no longer a build dependency
 * New concepts page explaining the different styles of neighbor lists
+* Default neighbor list buffer radius is more clearly shown to be r_buff = 0.4
 
 *Removed*